% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_dist.R
\name{age_dist}
\alias{age_dist}
\title{Stable age distribution}
\usage{
age_dist(Sf, Sm, r, pups)
}
\arguments{
\item{Sf}{a matrix of age by year survival rates for females}

\item{Sm}{a matrix of age by year survival rates for males}

\item{r}{assumed rate of increase at beginning of time series to create initial stable age distribution}

\item{pups}{number of pups born in the initial year}
}
\value{
list with abundance by age for males (N_m), females (N_f) and total N.
}
\description{
Computes sex-specific stable age distribution and abundance from pup count
from survival for females and males (Sf and Sm) and an assumed rate of growth.
The age proportions is the turned into abundance by computing N=pups/c(0) where c(0) 
is the proportion that are pups in the age distribution. Actually it is
pups/2/cf(0) + pups/2/cm(0) where cf(0) and cm(0) are the proportions that are pups for
females and males respectively. Then N is split into the sex and age categories based on
the stable age distribution.
}
\author{
Jeff Laake
}

