% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots.R
\name{barplot_pop}
\alias{barplot_pop}
\title{Population projection barplots}
\usage{
barplot_pop(zcpop, begin = 1975, range = NULL)
}
\arguments{
\item{zcpop}{population array of numbers at each sex, age and time}

\item{begin}{year to initiate population projection}

\item{range}{a range of years to be plotted}
}
\value{
list with cummulative female and male age structure by year
}
\description{
Plots population age structure over time
}
\author{
Jeff Laake
}

