% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pop}
\alias{plot_pop}
\title{Population projection plots}
\usage{
plot_pop(zcpop, begin = 1975, plot = TRUE, span = 1/3)
}
\arguments{
\item{zcpop}{population array of numbers at each sex, age and time}

\item{begin}{year to initiate population projection}

\item{plot}{if TRUE, produces plots; otherwie, only returns stats}
}
\value{
vector containing year of maximum smoothed net increase, smoothed population size at MNPL and current smoothed N, all for non-pups.
}
\description{
Computes popn statistics and optionally plots smoothed net change, pup counts, total population size, non-pup population size over time
}
\author{
Jeff Laake
}

