% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popn_reconstruction.r
\name{popn_reconstruction}
\alias{make_predictions}
\alias{make_pup_counts}
\alias{popn_reconstruction}
\alias{popn_reconstruction_error}
\alias{population_reconstruction}
\alias{population_reconstruction_error}
\title{Population projection plots}
\usage{
popn_reconstruction(pup_counts, model, Phiformula, Phidata, Phi_coef = NULL,
  random = FALSE, marked = FALSE, df = NULL, dfu = NULL, begin = 1975,
  plot = TRUE, r = 0.05, Phi_sigma = NULL)
}
\arguments{
\item{pup_counts}{vector of pup counts}

\item{model}{survival model object from marked}

\item{Phiformula}{formula for survival model}

\item{Phidata}{dataframe with values for survival predictions}

\item{Phi_coef}{fixed effect coefficients for Phi}

\item{random}{if TRUE, add random component for survival calculation in dfu$u}

\item{df}{portion of design data for survival for which calculatoins are to be performed with Phidata values}

\item{dfu}{matrix with normal random variate for each prediction}

\item{begin}{year to initiate population projection}

\item{plot}{if TRUE, project_pop produces plots; otherwise, only returns stats}

\item{r}{assumed rate of increase at beginning of time series to create initial stable age distribution}

\item{Phi_sigma}{vector of estimates for log(sigma)}

\item{nrep}{number of replicate simulations}
}
\value{
vector containing year of maximum smoothed net increase, smoothed population size at MNPL and current smoothed N, all for non-pups.
}
\description{
Computes popn statistics and optionally plots smoothed net change, pup counts, total population size, non-pup population size over time
}
\author{
Jeff Laake
}

