% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_forward.R
\name{project_forward}
\alias{project_forward}
\title{Forward Population projection}
\usage{
project_forward(zcpop, birth, nyears)
}
\arguments{
\item{birth}{assumed rate of birth (pups/ repro female)}

\item{nyears}{number of years forward to project}

\item{population}{array of numbers at each sex, age and time}
}
\value{
population array of numbers at each sex, age and time extended foward
}
\description{
Projects an expected population with a 25 year age structure forward from the endpoint of
a population array with an assumed birth rate of reproductive females. It uses survival rate 
from final year in projecting forward.
}
\author{
Jeff Laake
}

