% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project._pop.R
\name{project_pop}
\alias{project_pop}
\title{Population projection}
\usage{
project_pop(pup_counts, Sf, Sm, r = 0.05, begin = 1975, plot = TRUE,
  span = 1/3)
}
\arguments{
\item{pup_counts}{vector of annual pup counts}

\item{Sf}{a matrix of age by year survival rates for females}

\item{Sm}{a matrix of age by year survival rates for males}

\item{r}{assumed rate of increase at beginning of time series to create initial stable age distribution}

\item{begin}{year to initiate population projection}

\item{plot}{if TRUE, produces plots; otherwie, only returns stats}
}
\value{
list with population array of numbers at each sex, age and time and vector of statistics for MNPL etc
}
\description{
Projects a population with a 25 year age structure from a series of 
annual pup counts and a sex- and age-specific survival rate matrix.
Animals can reach age 24 and then they disappear (S from age 24 to 25 is assumed to be 0).
}
\author{
Jeff Laake
}

